+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## 1.6.0
 
-* added new hourly visitor statistics
+* added hourly visitor statistics, listing events, os and browser versions
+* added filter options
+* updated return types with new and modified fields
+* improved error messages
 
 ## 1.5.1